+++ packages/zentara/test/fixtures/flash/admin.ts
@@ -1,0 +1,5 @@
+import { HttpError } from "../../../src/core/index.js";
+
+export const GET = () => {
+  throw new HttpError(403, "Khusus admin <toko>");
+};